Changes with some tinkering with the jump code
The jump height and force was changed.
I altered my test scene a bit.
Other pieces of code were altered and experimented with.
ASHLYNN spawn is broken, I've been able to work with it but I didn't fully fix it this didn't stop me from debugging though but it slowed me down.